requirements.txt
igor==0.3
imgaug==0.4.0
ipython==8.12.0
matplotlib==3.7.1
numpy>=1.22
pandas==1.5.3
Pillow==9.4.0
scikit_image==0.19.3
scikit_learn==1.2.2
seaborn==0.12.2
setuptools==61.2.0
toml==0.10.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
jupyterlab==3.5.3
scipy>=1
ipywidgets>=8
bokeh==3.1.1